refactor:Refactor toda las estructuras en modulos mas ordenados
De manera que las estructuras y en general los tipos y simbolos estan solo en 4 archivos diferentes
