--- Structures/Tipos/Cadena.py
@@ -1,28 +1,0 @@
-from Structures.Tipos.Tipo import *
-
-class Cadena(Tipo):
-    def __init__(self, nombreTipo="cadena", valor="cadena") -> None:
-        super().__init__(nombreTipo, valor)
-    def __str__(self) -> str:
-        return "cadena"
-    # >
-    def mayorQue(val1, val2):
-        pass
-    # <
-    def menorQue(val1, val2):
-        pass
-    # >=
-    def mayorIgualQue(val1, val2):
-        pass
-    # <=
-    def menorIgualQue(val1, val2):
-        pass
-    # ==
-    def igualQue(val1, val2):
-        pass
-    # !=
-    def diferenteQue(val1, val2):
-        pass
-    # +
-    def suma(val1, val2):
-        pass
