--- Structures/Tipos/DefinidoPorUsuario.py
@@ -1,6 +1,0 @@
-from Tipo import *
-
-class DefinidoPorUsuario(Tipo):
-    def __init__(self, nombreTipo="tipo", valor="tipo") -> None:
-        super().__init__(nombreTipo, valor)
-
