--- Structures/Tipos/Nil.py
@@ -1,25 +1,0 @@
-from Structures.Tipos.Tipo import *
-
-class Nil(Tipo):
-    def __init__(self, nombreTipo="nil", valor="nil") -> None:
-        super().__init__(nombreTipo, valor)
-    def __str__(self) -> str:
-        return "nil"
-    # >
-    def mayorQue(val1, val2):
-        pass
-    # <
-    def menorQue(val1, val2):
-        pass
-    # >=
-    def mayorIgualQue(val1, val2):
-        pass
-    # <=
-    def menorIgualQue(val1, val2):
-        pass
-    # ==
-    def igualQue(val1, val2):
-        pass
-    # !=
-    def diferenteQue(val1, val2):
-        pass
