--- Structures/Tipos/TipoFuncion.py
@@ -1,9 +1,0 @@
-from .Tipo import *
-
-class TipoFuncion(Tipo):
-    def __init__(self, nombreTipo="tipo", valor="tipo") -> None:
-        super().__init__(nombreTipo, valor)
-        
-    def __str__(self) -> str:
-        return "funcion"    
-
